completed some screens, implemented security and removed unused classes

+++ src/main/java/com/volvo/service/Service1.java
@@ -43,6 +43,6 @@ public class Service1 implements Service1I{
 
     @Override
     public int count() {
-        return targetDAO.count();
+        return (int) targetDAO.count();
     }
 }
